Added '+' support.
I've added support for the '+' command.
It appends, rather than newlines things.

It wasn't quite as smooth as I expected, but that's only because the '+' character is a valid prefix for a number to have.
I worked around it a little. Not quite a hack, just some mostly inaccessable default behaviour that happens to follow '+'.