Flipped Escape to default, and Fixed bug with 'a' and escape.
I accidentally mixed two distinct changes here.
So ashamed.

I've changed escpaed to be the default mode.
This required changeing the flag to be '-r' for raw mode, which means "Don't Escape"

Also, when 'a' was used as the command, it did it's own thing, which didn't take escaping into account.
This is now fixed.